+++ NotifySDK/NotifyCore.framework/Headers/NotifyCore.h
@@ -2,6 +2,9 @@
 
 #import <NotifyCore/NTFApp.h>
 #import <NotifyCore/NTFAppConfig.h>
+#import <NotifyCore/NTFNotifyConfig.h>
+#import <NotifyCore/NTFNotifyActivitySettings.h>
+#import <NotifyCore/NTFNotifyActivityColorSettings.h>
 #import <NotifyCore/NTFConstants.h>
 #import <NotifyCore/NTFMacros.h>
 #import <NotifyCore/NTFUnhandledErrorReceiver.h>